# exp/orderbook: Improve performance of path finding implementation