@ire-and-curses 2021-08-11
Wow - that's a factor 12 speedup! :tada: